+++ include/GafferImageUI/ImageView.h
@@ -103,6 +103,10 @@ class GAFFERIMAGEUI_API ImageView : public GafferUI::View
 		Gaffer::BoolPlug *lutGPUPlug();
 		const Gaffer::BoolPlug *lutGPUPlug() const;
 
+		/// The gadget responsible for displaying the image.
+		ImageGadget *imageGadget();
+		const ImageGadget *imageGadget() const;
+
 		void setContext( Gaffer::ContextPtr context ) override;
 
 		typedef std::function<GafferImage::ImageProcessorPtr ()> DisplayTransformCreator;